backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
sqlalchemy>=2.0.36
pymysql>=1.1.1
cryptography>=44.0.0
pydantic>=2.10.0
python-dotenv>=1.0.1
alembic>=1.14.0
